pyproject.toml
[tool.poetry]
name = "hedjobs"
version = "0.1.0"
description = "Job hiring platform"
authors = ["hthe1010"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.2"
pydantic-settings = "^2.0.3"
uvicorn = { version = "^0.23.2", extras = ["standard"] }
python-dotenv = "^1.0.0"
SQLAlchemy = "^2.0.21"
psycopg2-binary = "^2.9.3"
alembic = "^1.12.0"
httpx = "^0.25.0"
pytest = "^7.4.2"
email-validator = "^2.1.0"
passlib = { version = "^1.7.4", extras = ["bcrypt"] }
fastapi-mail = "^1.4.1"
Jinja2 = "^3.1.2"
PyJWT = "^2.8.0"
python-multipart = "^0.0.6"
fastapi-sqlalchemy = "^0.2.1"
unidecode = "^1.3.8"
black = "^24.4.2"
pre-commit = "^3.7.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
